requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
dominate==2.8.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.4
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
phonenumbers==8.13.14
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.16
twilio==8.3.0
typing_extensions==4.6.3
urllib3==2.0.3
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
